More idiomatic way of placing players on systems.
This approach is easier to maintain and improve. The footprint of the
specific behavior of the "naive" placement approach is kept minimal in
order to remove it easily next time we want to implement a better
approach. Which we will have to if we want to add factions or other
better placement scheme.