Use iterator idiom for income gathering.
It is better for the compiler in order for it to optimize code.
Sometimes, readability is better with a for loop, but here I think it is
more convenient to use iterators.

Moreover, using `values()` instead of `iter()` enable us to optimize the
way we walk through the systems `HashMap`.